tests/wpt/web-platform-tests/app-history/navigate/navigate-base-url.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="resources/page-with-base-url-common.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.appHistory.navigate("blank.html");
    i.onload = t.step_func_done(() => {
      const iframeURL = new URL(i.contentWindow.appHistory.current.url);
      assert_equals(iframeURL.pathname, "/common/blank.html");
    });
  });
}, "navigate() must resolve URLs relative to appHistory object's base URL");
</script>
